backend/requirements.txt
Flask==2.1.0
plaid_python==14.0.0
python-dotenv==0.15.0
itsdangerous==2.0.1
Werkzeug==2.2.2
pyairtable==2.2.0
